requirements.txt
gdown
streamlit
tensorflow
pillow
numpy

